+++ pom.xml
@@ -21,7 +21,7 @@
   <groupId>com.google.cloud.spark.bigtable</groupId>
   <artifactId>spark-bigtable-connector</artifactId>
   <packaging>pom</packaging>
-  <version>0.4.0-SNAPSHOT</version>  <!-- ${NEXT_VERSION_FLAG} -->
+  <version>0.4.0</version>  <!-- ${NEXT_VERSION_FLAG} -->
   <name>Spark Bigtable Connector Build Parent</name>
   <description>Parent project for all the Spark Bigtable Connector artifacts</description>
   <url>https://github.com/GoogleCloudDataproc/spark-bigtable-connector</url>
@@ -51,7 +51,8 @@
   </scm>
 
   <modules>
-    <module>spark-bigtable</module>
+    <module>spark-bigtable_2.12</module>
+    <module>spark-bigtable_2.13</module>
     <module>spark-bigtable-core-it</module>
   </modules>
 
@@ -72,7 +73,7 @@
 
     <scala.version>2.13.14</scala.version>
     <scala.binary.version>2.13</scala.binary.version>
-    <connector.artifact.id>spark-bigtable-scala2.13</connector.artifact.id>
+    <connector.artifact.id>spark-bigtable_2.13</connector.artifact.id>
     <spark.version>3.5.1</spark.version>
     <bigtable.java.version>2.42.0</bigtable.java.version>
     <bigtable.java.emulator.version>0.175.0</bigtable.java.emulator.version>
